requirements.txt
asgiref==3.7.2
attrs==23.2.0
colorama==0.4.6
Django==5.0.3
django-autoslug==1.9.9
django-cities==0.6.1
django-cities-light==3.10.1
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-filter==24.1
django-js-asset==2.2.0
django-model-utils==4.4.0
django-modeltranslation==0.18.11
django-mptt==0.16.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.0
phonenumbers==8.13.28
phonenumberslite==8.13.28
pillow==10.2.0
progressbar2==4.4.1
PyJWT==1.7.1
python-utils==3.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
swapper==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
